move RunError to its own module that we'll reuse in the outstation
