+++ dnp3/src/lib.rs
@@ -69,5 +69,6 @@ pub mod serial;
 /// Entry points and types for TCP
 pub mod tcp;
 
+pub(crate) mod shared;
 pub(crate) mod transport;
 pub(crate) mod util;
