+++ dnp3/src/master/error.rs
@@ -6,13 +6,13 @@ use crate::app::{ObjectParseError, Shutdown};
 use crate::link::error::LinkError;
 use crate::link::EndpointAddress;
 use crate::master::association::NoAssociation;
-use crate::master::session::{RunError, StopReason};
 use crate::transport::TransportResponseError;
 
 use tokio::sync::mpsc::error::SendError;
 use tokio::sync::oneshot::error::RecvError;
 
 use crate::app::attr::BadAttribute;
+use crate::shared::{RunError, StopReason};
 
 /// Errors that can occur when adding an association
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
