+++ dnp3/src/shared/mod.rs
@@ -1,0 +1,2 @@
+mod errors;
+pub(crate) use errors::*;
