+++ dnp3/src/tcp/master.rs
@@ -4,8 +4,9 @@ use tracing::Instrument;
 use crate::app::{ConnectStrategy, Listener};
 use crate::app::{RetryStrategy, Shutdown};
 use crate::link::LinkErrorMode;
-use crate::master::session::{MasterSession, RunError, StopReason};
+use crate::master::session::MasterSession;
 use crate::master::{MasterChannel, MasterChannelConfig};
+use crate::shared::{RunError, StopReason};
 use crate::tcp::EndpointList;
 use crate::tcp::{ClientState, ConnectOptions, Connector, PostConnectionHandler};
 use crate::transport::TransportReader;
